requirements.txt
Flask
Jinja2
requests
slackclient
slackeventsapi
websocket-client
Werkzeug
